feat(client): make small and large version of SearchField
