+++ src/client/src/molecules/selection/__stories__/Button.stories.tsx
@@ -4,62 +4,64 @@ import getDisplayName from 'react-display-name';
 import { StorybookWrapper } from '../../../helpers/StorybookWrapper';
 import { Button, ButtonStyle } from '../Button';
 
-storiesOf('molecules/selection/Button/contained/primary', module).addWithJSX(
-  getDisplayName(Button),
-  () => (
-    <StorybookWrapper>
-      <Button buttonStyle={ButtonStyle.Contained}>sign in</Button>
-    </StorybookWrapper>
-  ),
-);
+const componentName = getDisplayName(Button);
 
-storiesOf('molecules/selection/Button/contained/secondary', module).addWithJSX(
-  getDisplayName(Button),
-  () => (
-    <StorybookWrapper>
-      <Button buttonStyle={ButtonStyle.Contained} secondary={true}>
-        sign in
-      </Button>
-    </StorybookWrapper>
-  ),
-);
+storiesOf(
+  `molecules/selection/${componentName}/contained/primary`,
+  module,
+).addWithJSX(componentName, () => (
+  <StorybookWrapper>
+    <Button buttonStyle={ButtonStyle.Contained}>sign in</Button>
+  </StorybookWrapper>
+));
 
-storiesOf('molecules/selection/Button/outlined/primary', module).addWithJSX(
-  getDisplayName(Button),
-  () => (
-    <StorybookWrapper>
-      <Button buttonStyle={ButtonStyle.Outlined}>sign in</Button>
-    </StorybookWrapper>
-  ),
-);
+storiesOf(
+  `molecules/selection/${componentName}/contained/secondary`,
+  module,
+).addWithJSX(componentName, () => (
+  <StorybookWrapper>
+    <Button buttonStyle={ButtonStyle.Contained} secondary={true}>
+      sign in
+    </Button>
+  </StorybookWrapper>
+));
 
-storiesOf('molecules/selection/Button/outlined/secondary', module).addWithJSX(
-  getDisplayName(Button),
-  () => (
-    <StorybookWrapper>
-      <Button buttonStyle={ButtonStyle.Outlined} secondary={true}>
-        sign in
-      </Button>
-    </StorybookWrapper>
-  ),
-);
+storiesOf(
+  `molecules/selection/${componentName}/outlined/primary`,
+  module,
+).addWithJSX(componentName, () => (
+  <StorybookWrapper>
+    <Button buttonStyle={ButtonStyle.Outlined}>sign in</Button>
+  </StorybookWrapper>
+));
 
-storiesOf('molecules/selection/Button/text/primary', module).addWithJSX(
-  getDisplayName(Button),
-  () => (
-    <StorybookWrapper>
-      <Button buttonStyle={ButtonStyle.Text}>sign in</Button>
-    </StorybookWrapper>
-  ),
-);
+storiesOf(
+  `molecules/selection/${componentName}/outlined/secondary`,
+  module,
+).addWithJSX(componentName, () => (
+  <StorybookWrapper>
+    <Button buttonStyle={ButtonStyle.Outlined} secondary={true}>
+      sign in
+    </Button>
+  </StorybookWrapper>
+));
 
-storiesOf('molecules/selection/Button/text/secondary', module).addWithJSX(
-  getDisplayName(Button),
-  () => (
-    <StorybookWrapper>
-      <Button buttonStyle={ButtonStyle.Text} secondary={true}>
-        sign in
-      </Button>
-    </StorybookWrapper>
-  ),
-);
+storiesOf(
+  `molecules/selection/${componentName}/text/primary`,
+  module,
+).addWithJSX(componentName, () => (
+  <StorybookWrapper>
+    <Button buttonStyle={ButtonStyle.Text}>sign in</Button>
+  </StorybookWrapper>
+));
+
+storiesOf(
+  `molecules/selection/${componentName}/text/secondary`,
+  module,
+).addWithJSX(componentName, () => (
+  <StorybookWrapper>
+    <Button buttonStyle={ButtonStyle.Text} secondary={true}>
+      sign in
+    </Button>
+  </StorybookWrapper>
+));
